+++ src/metatrain/share/base_hypers.py
@@ -448,8 +448,8 @@ class BaseHypers(TypedDict):
     """Configuration for Weights & Biases logging.
 
     If ``None``, W&B logging is disabled."""
-    evaluate_best_model: NotRequired[bool] = True
-    """Whether to evaluate the best model (the one exported by ``mtt train``) on the
+    final_eval: NotRequired[bool] = True
+    """Whether to evaluate the final model (the one exported by ``mtt train``) on the
     training, validation and test sets after training, logging the resulting RMSE/MAE
     metrics.